Implement `readInput()` as an async function on the front side
